Refactored Code to be cleaner
