+++ api/ragbot_tools/collate_book_info_for_vector_db_creation.py
@@ -14,7 +14,7 @@ from langchain.schema import Document
 
 # import supabase
 from supabase.client import Client, create_client
-from supabase_functions import get_all_records
+from api.tools.supabase_functions import get_all_records
 
 # load environment variables
 load_dotenv()
