+++ api/ragbot_tools/sql_to_create_vector_db.txt
