+++ api/test_files/test_isbn_function.py
@@ -1,7 +1,7 @@
 # should add  in more tests - CI/CD - PRs - GitHub Build
 import pytest
 
-from book_functions import get_isbn_for_book, get_google_books_details_using_isbn, get_book_meta_data_from_isbn, create_book_record
+from api.tools.book_functions import get_isbn_for_book, get_google_books_details_using_isbn, get_book_meta_data_from_isbn, create_book_record
 
 # title = "The Hitchhiker's Guide to the Galaxy"
 # author = "Douglas Adams"
