+++ api/tools/image_recognition.py
