+++ requirements.txt
@@ -26,4 +26,9 @@ pypdf==5.3.0
 
 scikit-learn==1.7.1
 
+flask-swagger-ui==5.21.0
+Flask-RESTful==0.3.10
+flask-swagger==0.2.14
+flasgger==0.9.7.1
+
 # Note: Ensure that Tesseract OCR is installed on your system for pytesseract to work.
